Fixed #21: Implemented a complete custom type to make sure the SPI matches expectations. That said some problems were found and the SPI and DSL are adjusted accordingly. A new DSL part for writing custom or complex semantic tags was added based on SemanticTagSupport#semanticTag and the magic capture based on Proxy

+++ src/main/java/com/noctarius/borabora/spi/TypeSpec.java
@@ -27,7 +27,7 @@ public interface TypeSpec {
 
     String spec();
 
-    int tagId();
+    long tagId();
 
     boolean matches(TypeSpec other);
